requirements.txt
asgiref==3.7.2
Django==4.2.3
Pillow==10.0.0
pypng==0.20220715.0
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.7.1
